Simplify networking config for catalyst
This commit removes the bridge interface and disables ipv6 on the
default networking config. This seems to be causing DNS issues with some
programs, mainly atm10. It may be conflicting with systemd-networkd
since that is primarily used for networking now. Disabling this still
results in link local ipv6 addresses being assigned to the interface,
which backs up this theory. This commit also disabled immich, since it
is the only service relying on the bridge, and was not working anyway.
Will need to revisit the immich config and reenable later.